Adding claimDate to EVSS payload (#3596)
* Adding claimDate to EVSS payload

* removing commented out libs

* quick code climate refactor

* fixed validate

* fixing claim spec

* fixing spec

* one more test

* rubocop fix
