+++ modules/claims_api/app/controllers/claims_api/v0/forms/disability_compensation_controller.rb
@@ -25,7 +25,7 @@ module ClaimsApi
             source: request.headers['X-Consumer-Username']
           )
           auto_claim = ClaimsApi::AutoEstablishedClaim.find_by(md5: auto_claim.md5) unless auto_claim.id
-          service_object.validate_form526(auto_claim.form.to_internal)
+          service_object.validate_form526(auto_claim.to_internal)
 
           ClaimsApi::ClaimEstablisher.perform_async(auto_claim.id)
 
@@ -48,17 +48,7 @@ module ClaimsApi
         end
 
         def validate_form_526
-          service = EVSS::DisabilityCompensationForm::ServiceAllClaim.new(auth_headers)
-          auto_claim = ClaimsApi::AutoEstablishedClaim.new(
-            status: ClaimsApi::AutoEstablishedClaim::PENDING,
-            auth_headers: auth_headers,
-            form_data: form_attributes
-          )
-          service.validate_form526(auto_claim.form.to_internal)
-          render json: valid_526_response
-        rescue EVSS::ErrorMiddleware::EVSSError => e
-          track_526_validation_errors(e.details)
-          render json: { errors: format_526_errors(e.details) }, status: :unprocessable_entity
+          super
         end
       end
     end
