+++ modules/claims_api/config/schemas/526.json
@@ -649,6 +649,9 @@
     },
     "autoCestPDFGenerationDisabled": {
       "type": "boolean"
+    },
+    "claimDate": {
+      "$ref": "#/definitions/date"
     }
   }
 }
