+++ modules/claims_api/spec/fixtures/form_526_json_api.json
@@ -33,6 +33,7 @@
       ],
       "standardClaim": false,
       "autoCestPDFGenerationDisabled": false,
+      "claimDate": "1990-01-03",
       "applicationExpirationDate": "2015-08-28T19:53:45+00:00",
       "serviceInformation": {
         "servicePeriods": [
